final2

+++ Binomial.java
@@ -20,8 +20,8 @@ public class Binomial {
 		return binomial1 (n -1,k) + binomial1 ( n-1 , k-1);
 	 }
 	
-	public static int binomial(int n, int k) {
-		int [][] memo = new int [n +1][k + 1];
+	public static long binomial(int n, int k) {
+		long [][] memo = new long [n +1][k + 1];
 		for (int i = 0; i <= n; i++) {
         for (int j = 0; j <= k; j++) {
           memo[i][j] = -1;
@@ -31,7 +31,7 @@ public class Binomial {
 		return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
